Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   README.md
	new file:   yolov3/detection.py
	modified:   yolov3/images/back0.jpg
	modified:   yolov3/images/back0.txt
	modified:   yolov3/images/back1.jpg
	modified:   yolov3/images/back1.txt
	modified:   yolov3/images/back10.jpg
	modified:   yolov3/images/back10.txt
	modified:   yolov3/images/back11.jpg
	modified:   yolov3/images/back11.txt
	modified:   yolov3/images/back12.jpg
	modified:   yolov3/images/back12.txt
	modified:   yolov3/images/back13.jpg
	modified:   yolov3/images/back13.txt
	modified:   yolov3/images/back14.jpg
	modified:   yolov3/images/back14.txt
	modified:   yolov3/images/back15.jpg
	modified:   yolov3/images/back15.txt
	modified:   yolov3/images/back16.jpg
	modified:   yolov3/images/back16.txt
	modified:   yolov3/images/back17.jpg
	modified:   yolov3/images/back17.txt
	modified:   yolov3/images/back18.jpg
	modified:   yolov3/images/back18.txt
	modified:   yolov3/images/back19.jpg
	modified:   yolov3/images/back19.txt
	modified:   yolov3/images/back2.jpg
	modified:   yolov3/images/back2.txt
	modified:   yolov3/images/back20.jpg
	modified:   yolov3/images/back20.txt
	modified:   yolov3/images/back21.jpg
	modified:   yolov3/images/back21.txt
	modified:   yolov3/images/back22.jpg
	modified:   yolov3/images/back22.txt
	modified:   yolov3/images/back23.jpg
	modified:   yolov3/images/back23.txt
	modified:   yolov3/images/back24.jpg
	modified:   yolov3/images/back24.txt
	modified:   yolov3/images/back25.jpg
	modified:   yolov3/images/back25.txt
	modified:   yolov3/images/back26.jpg
	modified:   yolov3/images/back26.txt
	modified:   yolov3/images/back27.jpg
	modified:   yolov3/images/back27.txt
	modified:   yolov3/images/back28.jpg
	modified:   yolov3/images/back28.txt
	modified:   yolov3/images/back29.jpg
	modified:   yolov3/images/back29.txt
	modified:   yolov3/images/back3.jpg
	modified:   yolov3/images/back3.txt
	modified:   yolov3/images/back30.jpg
	modified:   yolov3/images/back30.txt
	modified:   yolov3/images/back31.jpg
	modified:   yolov3/images/back31.txt
	modified:   yolov3/images/back32.jpg
	modified:   yolov3/images/back32.txt
	modified:   yolov3/images/back33.jpg
	modified:   yolov3/images/back33.txt
	modified:   yolov3/images/back34.jpg
	modified:   yolov3/images/back34.txt
	modified:   yolov3/images/back35.jpg
	modified:   yolov3/images/back35.txt
	modified:   yolov3/images/back36.jpg
	modified:   yolov3/images/back36.txt
	modified:   yolov3/images/back37.jpg
	modified:   yolov3/images/back37.txt
	modified:   yolov3/images/back38.jpg
	modified:   yolov3/images/back38.txt
	modified:   yolov3/images/back39.jpg
	modified:   yolov3/images/back39.txt
	modified:   yolov3/images/back4.jpg
	modified:   yolov3/images/back4.txt
	modified:   yolov3/images/back40.jpg
	modified:   yolov3/images/back40.txt
	modified:   yolov3/images/back41.jpg
	modified:   yolov3/images/back41.txt
	modified:   yolov3/images/back42.jpg
	modified:   yolov3/images/back42.txt
	modified:   yolov3/images/back43.jpg
	modified:   yolov3/images/back43.txt
	modified:   yolov3/images/back44.jpg
	modified:   yolov3/images/back44.txt
	modified:   yolov3/images/back45.jpg
	modified:   yolov3/images/back45.txt
	modified:   yolov3/images/back46.jpg
	modified:   yolov3/images/back46.txt
	modified:   yolov3/images/back47.jpg
	modified:   yolov3/images/back47.txt
	modified:   yolov3/images/back48.jpg
	modified:   yolov3/images/back48.txt
	modified:   yolov3/images/back49.jpg
	modified:   yolov3/images/back49.txt
	modified:   yolov3/images/back5.jpg
	modified:   yolov3/images/back5.txt
	modified:   yolov3/images/back50.jpg
	modified:   yolov3/images/back50.txt
	modified:   yolov3/images/back51.jpg
	modified:   yolov3/images/back51.txt
	modified:   yolov3/images/back52.jpg
	modified:   yolov3/images/back52.txt
	modified:   yolov3/images/back53.jpg
	modified:   yolov3/images/back53.txt
	modified:   yolov3/images/back54.jpg
	modified:   yolov3/images/back54.txt
	modified:   yolov3/images/back55.jpg
	modified:   yolov3/images/back55.txt
	modified:   yolov3/images/back56.jpg
	modified:   yolov3/images/back56.txt
	modified:   yolov3/images/back57.jpg
	modified:   yolov3/images/back57.txt
	modified:   yolov3/images/back58.jpg
	modified:   yolov3/images/back58.txt
	modified:   yolov3/images/back59.jpg
	modified:   yolov3/images/back59.txt
	modified:   yolov3/images/back6.jpg
	modified:   yolov3/images/back6.txt
	modified:   yolov3/images/back60.jpg
	modified:   yolov3/images/back60.txt
	modified:   yolov3/images/back61.jpg
	modified:   yolov3/images/back61.txt
	modified:   yolov3/images/back62.jpg
	modified:   yolov3/images/back62.txt
	modified:   yolov3/images/back63.jpg
	modified:   yolov3/images/back63.txt
	modified:   yolov3/images/back64.jpg
	modified:   yolov3/images/back64.txt
	modified:   yolov3/images/back65.jpg
	modified:   yolov3/images/back65.txt
	modified:   yolov3/images/back66.jpg
	modified:   yolov3/images/back66.txt
	modified:   yolov3/images/back67.jpg
	modified:   yolov3/images/back67.txt
	modified:   yolov3/images/back68.jpg
	modified:   yolov3/images/back68.txt
	modified:   yolov3/images/back69.jpg
	modified:   yolov3/images/back69.txt
	modified:   yolov3/images/back7.jpg
	modified:   yolov3/images/back7.txt
	new file:   yolov3/images/back70.jpg
	new file:   yolov3/images/back70.txt
	new file:   yolov3/images/back71.jpg
	new file:   yolov3/images/back71.txt
	new file:   yolov3/images/back72.jpg
	new file:   yolov3/images/back72.txt
	new file:   yolov3/images/back73.jpg
	new file:   yolov3/images/back73.txt
	new file:   yolov3/images/back74.jpg
	new file:   yolov3/images/back74.txt
	new file:   yolov3/images/back75.jpg
	new file:   yolov3/images/back75.txt
	new file:   yolov3/images/back76.jpg
	new file:   yolov3/images/back76.txt
	new file:   yolov3/images/back77.jpg
	new file:   yolov3/images/back77.txt
	new file:   yolov3/images/back78.jpg
	new file:   yolov3/images/back78.txt
	new file:   yolov3/images/back79.jpg
	new file:   yolov3/images/back79.txt
	modified:   yolov3/images/back8.jpg
	modified:   yolov3/images/back8.txt
	new file:   yolov3/images/back80.jpg
	new file:   yolov3/images/back80.txt
	new file:   yolov3/images/back81.jpg
	new file:   yolov3/images/back81.txt
	new file:   yolov3/images/back82.jpg
	new file:   yolov3/images/back82.txt
	new file:   yolov3/images/back83.jpg
	new file:   yolov3/images/back83.txt
	new file:   yolov3/images/back84.jpg
	new file:   yolov3/images/back84.txt
	new file:   yolov3/images/back85.jpg
	new file:   yolov3/images/back85.txt
	new file:   yolov3/images/back86.jpg
	new file:   yolov3/images/back86.txt
	new file:   yolov3/images/back87.jpg
	new file:   yolov3/images/back87.txt
	new file:   yolov3/images/back88.jpg
	new file:   yolov3/images/back88.txt
	new file:   yolov3/images/back89.jpg
	new file:   yolov3/images/back89.txt
	modified:   yolov3/images/back9.jpg
	modified:   yolov3/images/back9.txt
	new file:   yolov3/images/back90.jpg
	new file:   yolov3/images/back90.txt
	new file:   yolov3/images/back91.jpg
	new file:   yolov3/images/back91.txt
	new file:   yolov3/images/back92.jpg
	new file:   yolov3/images/back92.txt
	new file:   yolov3/images/back93.jpg
	new file:   yolov3/images/back93.txt
	new file:   yolov3/images/back94.jpg
	new file:   yolov3/images/back94.txt
	new file:   yolov3/images/back95.jpg
	new file:   yolov3/images/back95.txt
	new file:   yolov3/images/back96.jpg
	new file:   yolov3/images/back96.txt
	new file:   yolov3/images/back97.jpg
	new file:   yolov3/images/back97.txt
	new file:   yolov3/images/back98.jpg
	new file:   yolov3/images/back98.txt
	new file:   yolov3/images/back99.jpg
	new file:   yolov3/images/back99.txt
	new file:   yolov3/images/images.zip
	modified:   yolov3/rename.py

+++ yolov3/rename.py
@@ -5,17 +5,18 @@ FOLDER = './yolov3/'
 IMG_FOLDER = FOLDER + 'images/'
 TEMP_FOLDER = FOLDER + 'temp/'
 NAME = 'back'
-EXT = '.jpg'
+JPG_EXT = '.jpg'
+IMG_EXT = ['jpg', 'png']
 
 # Rename image files and txt files in sequential order 
 maxN = 0
 for i, fullName in enumerate(os.listdir(IMG_FOLDER)):
-    name, ext = fullName.split('.')
-    if ext != 'txt':
+    name, ext = fullName.rsplit('.', 1)
+    if ext in IMG_EXT:
         old_txtName = IMG_FOLDER + name + '.txt'
         new_txtName = TEMP_FOLDER + NAME + str(maxN) + '.txt'
         oldName = IMG_FOLDER + fullName
-        newName = TEMP_FOLDER + NAME + str(maxN) + EXT
+        newName = TEMP_FOLDER + NAME + str(maxN) + JPG_EXT
         maxN += 1
         os.rename(oldName, newName)
         if os.path.exists(old_txtName):
@@ -24,7 +25,7 @@ for i, fullName in enumerate(os.listdir(IMG_FOLDER)):
 # Delete file .txt without classes.txt
 for i, fullName in enumerate(os.listdir(IMG_FOLDER)):
     name, ext = fullName.split('.')
-    if name != 'classes':
+    if name != 'classes' and ext == 'txt':
         fileName = IMG_FOLDER + fullName
         os.remove(fileName)
